VMW 3D: Rebuild Mesa to include svga on aarch64
"svga" is the VMware graphics driver (vmwgfx). The default Mesa build
doesn't include this on aarch64[1] and Nix defers to the Mesa
defaults[2]. We override the Mesa build to explicitly include the "svga"
driver.

This makes 3d acceleration work on VMware Fusion 22H2! The "glxinfo"
still shows "Accelerated: no" but this appears to be a lie, since the
graphics performance is very noticeably better. "glxgears" is 10x faster
than llvmpipe.

[1]: https://github.com/Mesa3D/mesa/blob/49efa73ba11c4cacaed0052b984e1fb884cf7600/meson.build#L192
[2]: https://github.com/NixOS/nixpkgs/blob/71617faa487ab0d16b6cad3c20b8c4c79fb328e3/pkgs/development/libraries/mesa/default.nix#L9